refactor: add Poetry build system and tooling
