+++ .editorconfig
@@ -1,0 +1,37 @@
+# http://editorconfig.org
+
+# A special property that should be specified at the top of the file outside of any sections.
+# Set to true to stop .editorconfig file search on the current file.
+root                        = true
+
+[*]
+charset                     = utf-8
+trim_trailing_whitespace    = true
+insert_final_newline        = true
+indent_style                = space
+indent_size                 = 2
+continuation_indent_size    = 2
+max_line_length             = 120
+
+[*.sh]
+end_of_line                 = lf
+
+# shfmt options, equivalent to: -ln=bash -bn -ci -sr
+shell_variant               = bash
+binary_next_line            = true
+switch_case_indent          = true
+space_redirects             = true
+
+[*.{bat,cmd,ps1}]
+end_of_line                 = crlf
+
+[*.py]
+indent_size                 = 4
+continuation_indent_size    = 4
+
+[Makefile]
+indent_style                = tab
+end_of_line                 = lf
+
+[.gitmodules]
+indent_style = tab
