+++ .gitignore
@@ -1,0 +1,3 @@
+/dist/
+/poetry.lock
+__pycache__/
