--- setup.py
@@ -1,7 +1,0 @@
-from setuptools import setup
-setup(
-    name='fixsamplemessages',
-    version='1.0.1',
-    py_modules=['fixsamplemessages'],
-    install_requires = ['markdown>=3.0'],
-)
